# ~5% performance improvement by optimising main hot loop (TIMER_Clock)