Hoping I do not mess up this attempt at a PR.

The aim is to perform as little change as possible to improve performance.

The improvement idea is to optimise the very hot TIMER_Clock function by only processing cycles with events. Open to any feedback e.g. the variables are naming is very naiive

This results in an approximately 5% improvement as measured on macOS with clang (don't have access to other platforms)

### Test Results

Please note both the "took" and "done in" timings are provided from the command line as I am unsure which is more relevant 

**BEFORE**
| Item             | Run 1 | Run 2 | Run 3 | Run 4 | Run 5 | Mean   |
|------------------|-------|-------|-------|-------|-------|--------|
| 01.mid - TOOK    | 41.5  | 41.57 | 41.52 | 41.52 | 41.53 | 41.528 |
| 01.mid - DONE IN | 45.82 | 45.8  | 45.8  | 45.8  | 45.81 | 45.806 |
| 02.mid - TOOK    | 58.77 | 58.79 | 58.75 | 58.63 | 58.62 | 58.712 |
| 02.mid - DONE IN | 62.86 | 62.87 | 62.88 | 62.87 | 62.92 | 62.88  |

**AFTER**
| Item             | Run 1 | Run 2 | Run 3 | Run 4 | Run 5 | Mean   | Improvement |
|------------------|-------|-------|-------|-------|-------|--------|--------|
| 01.mid - TOOK    | 38.79 | 38.67 | 38.71 | 38.69 | 38.65 | 38.702 | 6.8%   |
| 01.mid - DONE IN | 42.58 | 42.59 | 42.59 | 42.59 | 42.59 | 42.588 | 7.0%   |
| 02.mid - TOOK    | 54.58 | 54.44 | 54.5  | 54.46 | 54.39 | 54.474 | 7.2%   |
| 02.mid - DONE IN | 58.68 | 58.67 | 58.66 | 58.67 | 58.67 | 58.67  | 6.7%   |



### Test method

5 x executions of 01.mid and 02.mid both pre and post changes. Command line used
`./Release/bin/nuked-sc55-render  --rom-directory ./_ROMS ../test/integration/avmidi/01.mid -r gm -f f32 --stdout --debug > /dev/null `


### Environment

macOS Version 15.3.2
Macbook pro Apple M1 Pro
Apple clang version 16.0.0 (clang-1600.0.26.6)